+++ libs/components/indicators/projects/indicators/src/modules/icon/fixtures/icon.component.fixture.ts
@@ -1,0 +1,19 @@
+import {
+  Component
+} from '@angular/core';
+
+import {
+  SkyIconVariantType
+} from '../types/icon-variant-type';
+
+@Component({
+  selector: 'sky-test-cmp',
+  templateUrl: './icon.component.fixture.html'
+})
+export class IconTestComponent {
+  public icon = 'circle';
+  public iconType: 'fa' | 'skyux';
+  public size = '3x';
+  public fixedWidth = false;
+  public variant: SkyIconVariantType;
+}