+++ libs/components/indicators/projects/indicators/src/modules/icon/icon-variant.ts
@@ -1,0 +1,16 @@
+/**
+ * @internal
+ * @deprecated Use `SkyIconVariantType` instead.
+ * Indicates the icon variant to display.
+ */
+export const enum SkyIconVariant {
+  /**
+   * Displays the line variant.
+   */
+  Line = 'line',
+
+  /**
+   * Displays the solid variant.
+   */
+  Solid = 'solid'
+}