+++ libs/components/indicators/projects/indicators/src/modules/label/fixtures/label.component.fixture.ts
@@ -1,0 +1,15 @@
+import {
+  Component
+} from '@angular/core';
+
+import {
+  SkyLabelType
+} from '../label-type';
+
+@Component({
+  selector: 'sky-test-cmp',
+  templateUrl: './label.component.fixture.html'
+})
+export class LabelTestComponent {
+  public labelType: SkyLabelType;
+}