+++ libs/components/indicators/projects/indicators/src/modules/text-highlight/text-highlight.directive.ts
@@ -1,0 +1,172 @@
+import {
+  AfterViewInit,
+  Directive,
+  ElementRef,
+  Input,
+  OnChanges,
+  OnDestroy,
+  SimpleChanges
+} from '@angular/core';
+
+import {
+  MutationObserverService
+} from '@skyux/core';
+
+const className = 'sky-highlight-mark';
+
+// Need to add the following to classes which contain static methods.
+// See: https://github.com/ng-packagr/ng-packagr/issues/641
+// @dynamic
+/**
+ * Highlights all matching text within DOM elements that it is placed on.
+ * The directive is only valid for elements that contain content, so you
+ * cannot place it on `ng-template`, `ng-content`, or `ng-container` tags.
+ */
+@Directive({
+  selector: '[skyHighlight]'
+})
+export class SkyTextHighlightDirective implements OnChanges, AfterViewInit, OnDestroy {
+
+  /**
+   * Specifies the text to highlight.
+   */
+  @Input()
+  public skyHighlight: string;
+
+  private existingHighlight = false;
+
+  private observer: MutationObserver;
+
+  constructor(
+    private el: ElementRef,
+    private observerService: MutationObserverService
+  ) { }
+
+  public ngOnChanges(changes: SimpleChanges): void {
+    if (changes.skyHighlight && !changes.skyHighlight.firstChange) {
+      this.highlight();
+    }
+  }
+
+  public ngAfterViewInit(): void {
+    const self = this;
+    this.observer = this.observerService.create((mutations: MutationRecord[]) => {
+      self.highlight();
+    });
+
+    this.observeDom();
+    if (this.skyHighlight) {
+      this.highlight();
+    }
+  }
+
+  public ngOnDestroy(): void {
+    /* istanbul ignore else */
+    if (this.observer) {
+      this.observer.disconnect();
+    }
+  }
+
+  private readyForHighlight(searchText: string): boolean {
+    return searchText && this.el.nativeElement;
+  }
+
+  private highlight(): void {
+    /* istanbul ignore else */
+    if (this.observer) {
+      this.observer.disconnect();
+    }
+
+    const searchText = this.skyHighlight;
+
+    if (this.existingHighlight) {
+      SkyTextHighlightDirective.removeHighlight(this.el);
+    }
+
+    /* istanbul ignore else */
+    if (this.readyForHighlight(searchText)) {
+      const node: HTMLElement = this.el.nativeElement;
+
+      // mark all matched text in the DOM
+      SkyTextHighlightDirective.markTextNodes(node, searchText);
+      this.existingHighlight = true;
+    }
+
+    this.observeDom();
+  }
+
+  private observeDom(): void {
+    /* istanbul ignore else */
+    if (this.observer) {
+      const config = { attributes: false, childList: true, characterData: true };
+      this.observer.observe(this.el.nativeElement, config);
+    }
+  }
+
+  private static cleanRegex(regex: string): string {
+    return regex.replace(/[-\/\\^$*+?.()|[\]{}]/g, '\\$&');
+  }
+
+  private static getRegexMatch(node: HTMLElement, searchText: string): RegExpExecArray {
+    const text = node.nodeValue;
+    const newSearchText = this.cleanRegex(searchText);
+    const searchRegex = new RegExp(newSearchText, 'gi');
+
+    return searchRegex.exec(text);
+  }
+
+  private static markNode(node: any, searchText: string): number {
+    const regexMatch = SkyTextHighlightDirective.getRegexMatch(node, searchText);
+
+    // found match
+    if (regexMatch && regexMatch.length > 0) {
+
+      // split apart text node with mark tags in the middle on the search term
+      const matchIndex = regexMatch.index;
+
+      const middle = node.splitText(matchIndex);
+      middle.splitText(searchText.length);
+      const middleClone = middle.cloneNode(true);
+
+      const markNode = document.createElement('mark');
+      markNode.className = className;
+      markNode.appendChild(middleClone);
+      middle.parentNode.replaceChild(markNode, middle);
+
+      return 1;
+    } else {
+      return 0;
+    }
+  }
+
+  private static markTextNodes(node: HTMLElement, searchText: string): number {
+    if (node.nodeType === 3) {
+      return SkyTextHighlightDirective.markNode(node, searchText);
+
+    } else if (node.nodeType === 1 && node.childNodes) {
+      for (let i = 0; i < node.childNodes.length; i++) {
+        const childNode = node.childNodes[i] as HTMLElement;
+        i += SkyTextHighlightDirective.markTextNodes(childNode, searchText);
+      }
+    }
+
+    return 0;
+  }
+
+  private static removeHighlight(el: ElementRef): void {
+    const matchedElements =
+      el.nativeElement.querySelectorAll(`mark.${className}`) as NodeList;
+
+    /* istanbul ignore else */
+    /* sanity check */
+    if (matchedElements) {
+      for (let i = 0; i < matchedElements.length; i++) {
+        const node = matchedElements[i];
+        const parentNode = node.parentNode;
+
+        parentNode.replaceChild(node.firstChild, node);
+        parentNode.normalize();
+      }
+    }
+  }
+}