+++ libs/components/indicators/src/app/docs/status-indicator/status-indicator-docs.component.ts
@@ -1,9 +1,0 @@
-import {
-  Component
-} from '@angular/core';
-
-@Component({
-  selector: 'app-status-indicator-docs',
-  templateUrl: './status-indicator-docs.component.html'
-})
-export class StatusIndicatorDocsComponent {}