+++ libs/components/indicators/src/app/visual/icon/icon-demo.component.ts
@@ -1,9 +1,0 @@
-import {
-  Component
-} from '@angular/core';
-
-@Component({
-  selector: 'sky-icon-demo',
-  templateUrl: './icon-demo.component.html'
-})
-export class SkyIconDemoComponent { }